feat: disable registration and matches

+++ server/src/main/kotlin/delta/codecharacter/server/match/MatchService.kt
@@ -34,6 +34,7 @@ import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.amqp.rabbit.annotation.RabbitListener
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.beans.factory.annotation.Value
 import org.springframework.http.HttpStatus
 import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder
 import org.springframework.messaging.simp.SimpMessagingTemplate
@@ -65,6 +66,7 @@ class MatchService(
     @Autowired private val autoMatchRepository: AutoMatchRepository
 ) {
     private var mapper: ObjectMapper = jackson2ObjectMapperBuilder.build()
+    @Value("\${environment.event-open}") private var eventOpen = true
     private val logger: Logger = LoggerFactory.getLogger(MatchService::class.java)
 
     private fun createSelfMatch(userId: UUID, codeRevisionId: UUID?, mapRevisionId: UUID?) {
@@ -153,6 +155,9 @@ class MatchService(
     }
 
     fun createDCMatch(userId: UUID, dailyChallengeMatchRequestDto: DailyChallengeMatchRequestDto) {
+        if (!eventOpen) {
+            throw CustomException(HttpStatus.BAD_REQUEST, "Invalid match request")
+        }
         val (_, chall, challType, _, completionStatus) =
             dailyChallengeService.getDailyChallengeByDateForUser(userId, true)
         if (completionStatus != null && completionStatus) {
@@ -193,6 +198,9 @@ class MatchService(
         gameService.sendGameRequest(game, code, language, map)
     }
     fun createMatch(userId: UUID, createMatchRequestDto: CreateMatchRequestDto) {
+        if (!eventOpen) {
+            throw CustomException(HttpStatus.BAD_REQUEST, "Invalid match request")
+        }
         when (createMatchRequestDto.mode) {
             MatchModeDto.SELF -> {
                 val (_, _, mapRevisionId, codeRevisionId) = createMatchRequestDto
@@ -211,17 +219,19 @@ class MatchService(
     }
 
     fun createAutoMatch() {
-        val topNUsers = publicUserService.getTopNUsers()
-        val userIds = topNUsers.map { it.userId }
-        val usernames = topNUsers.map { it.username }
-        logger.info("Auto matches started for users: $usernames")
-        autoMatchRepository.deleteAll()
-        userIds.forEachIndexed { i, userId ->
-            run {
-                for (j in i + 1 until userIds.size) {
-                    val opponentUsername = usernames[j]
-                    val matchId = createDualMatch(userId, opponentUsername, MatchModeEnum.AUTO)
-                    autoMatchRepository.save(AutoMatchEntity(matchId, 0))
+        if (eventOpen) {
+            val topNUsers = publicUserService.getTopNUsers()
+            val userIds = topNUsers.map { it.userId }
+            val usernames = topNUsers.map { it.username }
+            logger.info("Auto matches started for users: $usernames")
+            autoMatchRepository.deleteAll()
+            userIds.forEachIndexed { i, userId ->
+                run {
+                    for (j in i + 1 until userIds.size) {
+                        val opponentUsername = usernames[j]
+                        val matchId = createDualMatch(userId, opponentUsername, MatchModeEnum.AUTO)
+                        autoMatchRepository.save(AutoMatchEntity(matchId, 0))
+                    }
                 }
             }
         }
